+++ cubi_sak/snappy/__init__.py
@@ -31,6 +31,7 @@ from .itransfer_raw_data import setup_argparse as setup_argparse_itransfer_raw_d
 from .itransfer_ngs_mapping import setup_argparse as setup_argparse_itransfer_ngs_mapping
 from .itransfer_variant_calling import setup_argparse as setup_argparse_itransfer_variant_calling
 from .pull_sheet import setup_argparse as setup_argparse_pull_sheet
+from .kickoff import setup_argparse as setup_argparse_kickoff
 
 
 def setup_argparse(parser: argparse.ArgumentParser) -> None:
@@ -61,6 +62,8 @@ def setup_argparse(parser: argparse.ArgumentParser) -> None:
     setup_argparse_pull_sheet(
         subparsers.add_parser("pull-sheet", help="Pull SODAR sample sheet into biomedsheet")
     )
+
+    setup_argparse_kickoff(subparsers.add_parser("kickoff", help="Kick-off SNAPPY pipeline steps."))
 
 
 def run(args, parser, subparser):