+++ HISTORY.md
@@ -16,6 +16,7 @@
 - Adding `cubi-sak isa-tab validate` command.
 - Adding `cubi-sak isa-tab resolve-hpo` command.
 - Adding `cubi-sak sodar download-sheet` command.
+- Adding `cubi-sak snappy kickoff` command.
 
 ## v0.1.0